`use Test::Requires::Git ()` is useless, so add an explicit -nocheck option
`use Test::Requires::Git ()` prevents import() to be run, and so the
test_requires_git function is not exported.

Given the small number of cases when compile-time check is not desirable
(e.g. when testing Test::Requires::Git itself), we add an explicit option
to skip the compile-time checks.

It only makes sense as the sole option, so only check for that case.