Print a diagnostic with the git version number
This will add a data point (the git version that was used for the
tests) in the test reports.

The main issue is that it's going to print that for every test that
uses Test::Requires::Git, which makes it a little too verbose for
some projects.

This is why I also added a -quiet option to import(), so that it
can be silenced. This is not very satisfying, though.